pyproject.toml
[project]
name = "genai-aura-getting-started"
version = "0.1.0"
description = "This repo contains code to get a GenAI application started quickly with Neo4j Aura."
authors = [
    {name = "alex",email = "alexander.gilmore@neo4j.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=0.3.24,<0.4.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "langchain-neo4j (>=0.4.0,<0.5.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph (>=0.3.34,<0.4.0)",
    "streamlit (>=1.44.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
